Revert "feat: support reading tables from main data source (#7238)"
This reverts commit d1736b97686b688ff780076f4d10c55337da61b3.
