+++ packages/core/client/src/locale/zh-CN.json
@@ -1123,8 +1123,8 @@
   "Controls whether the current document is allowed to use the Web MIDI API. When this policy is disabled, the Promise returned by Navigator.requestMIDIAccess() will reject with a SecurityError DOMException.": "控制是否允许当前文档使用 Web MIDI API。当这种规则被禁用时，Navigator.requestMIDIAccess() 返回的 Promise 会拒绝并抛出错误 DOMException。",
   "Controls whether the current document is allowed to use the Payment Request API. When this policy is enabled, the PaymentRequest() constructor will throw a SecurityError DOMException.": "控制是否允许当前文档使用 Payment Request API。当这种规则被启用时，构造函数 PaymentRequest() 会抛出错误 SecurityError。",
   "After successful submission, the selected data blocks will be automatically refreshed.": "提交成功后，会自动刷新这里选中的数据区块。",
-  "Block Linkage rules": "区块联动规则",
-  "Field Linkage rules": "字段联动规则",
+  "Block Linkage rules":"区块联动规则",
+  "Field Linkage rules":"字段联动规则",
   "Reset link expiration": "重置链接有效期",
   "Exact day": "指定日期",
   "This Week": "本周",
@@ -1141,19 +1141,11 @@
   "Next Year": "明年",
   "Past": "过去",
   "Next": "未来",
-  "Calendar week": "日历周",
-  "Calendar Month": "日历月",
-  "Calendar Year": "日历年",
-  "Scan to input": "扫码录入",
-  "Date variables(Deprecated)":"日期变量(废弃)",
-  "Disable manual input": "禁止手动输入",
-  "Enable Scan": "启用扫码录入",
-  "Available Collections": "可选数据表",
-  "Selected Collections": "已选数据表",
-  "Search collections...": "搜索数据表...",
-  "items": "项",
-  "item": "项",
-  "Load collections": "加载数据表",
-  "Are you sure you want to load {{count}} collection(s)?": "你确定要加载 {{count}} 个数据表吗？",
-  "Confirm Load Collections": "确认加载数据表"
+  "Calendar week":"日历周",
+  "Calendar Month":"日历月",
+  "Calendar Year":"日历年",
+  "Scan to input":"扫码录入",
+  "Disable manual input":"禁止手动输入",
+  "Enable Scan":"启用扫码录入",
+  "Date variables(Deprecated)":"日期变量(废弃)"
 }
