+++ packages/core/data-source-manager/src/index.ts
@@ -17,4 +17,3 @@ export * from './load-default-actions';
 export * from './types';
 
 export * from './utils';
-export * from './interface-type-map';
