+++ packages/plugins/@nocobase/plugin-data-source-manager/src/client/component/MainDataSourceManager/Configuration/ConfigurationTable.tsx
@@ -16,8 +16,6 @@ import {
   DataSourceContext_deprecated,
   EditSubFieldAction,
   FieldSummary,
-  LoadCollection,
-  ResourceActionContext,
   SchemaComponent,
   SchemaComponentContext,
   TemplateSummary,
@@ -34,7 +32,6 @@ import React, { useContext, useMemo, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { CollectionFields } from './CollectionFields';
 import { collectionSchema } from './schemas/collections';
-import { useDataSourceRefresh } from '../../../hooks';
 
 /**
  * @param service
@@ -117,16 +114,6 @@ export const ConfigurationTable = () => {
   const compile = useCompile();
   const form = useForm();
   const app = useApp();
-  const service = useContext(ResourceActionContext);
-
-  const useRefreshActionProps = () => {
-    return useDataSourceRefresh({
-      dataSourceName: 'main',
-      onSuccess: () => {
-        service?.refresh?.();
-      },
-    });
-  };
 
   /**
    *
@@ -237,7 +224,6 @@ export const ConfigurationTable = () => {
           FieldSummary,
           TemplateSummay: TemplateSummary,
           CollectionFields,
-          LoadCollection,
         }}
         scope={{
           loadFilterTargetKeys,
@@ -256,7 +242,6 @@ export const ConfigurationTable = () => {
           enableInherits: database?.dialect === 'postgres',
           isPG: database?.dialect === 'postgres',
           getPickerFormat,
-          useRefreshActionProps,
         }}
       />
     </SchemaComponentContext.Provider>
