+++ packages/plugins/@nocobase/plugin-data-source-manager/src/client/hooks/index.ts
@@ -7,10 +7,6 @@
  * For more information, please refer to: https://www.nocobase.com/agreement.
  */
 
-export * from './useDataSourceRefresh';
-export * from './useResourceData';
-export * from './useDataSourceActions';
-
 import { useForm, useField } from '@formily/react';
 import { message } from 'antd';
 import { useTranslation } from 'react-i18next';
