+++ packages/plugins/@nocobase/plugin-data-source-manager/src/client/schema/index.ts
@@ -150,7 +150,6 @@ export const databaseConnectionSchema: ISchema = {
               type: 'checkbox',
             },
             useDataSource: '{{ cm.useDataSourceFromRAC }}',
-            onSuccess: '{{ dataSourceListCallback }}',
           },
           properties: {
             key: {
